Adding claimDate to EVSS payload
