+++ lib/evss/disability_compensation_form/data_translation.rb
@@ -19,6 +19,7 @@ module EVSS
       #
       def translate
         form['claimantCertification'] = true
+        # form['claimDate'] ||= application_create_date.try(:to_date).try(:to_s)
         form['autoCestPDFGenerationDisabled'] ||= false
         form['applicationExpirationDate'] = application_expiration_date
