+++ modules/claims_api/spec/fixtures/invalid_form_526_json_api.json
@@ -26,6 +26,7 @@
       ],
       "standardClaim": false,
       "autoCestPDFGenerationDisabled":false,
+      "claimDate": "1990-01-03",
       "applicationExpirationDate": "2015-08-28T19:53:45+00:00",
       "serviceInformation": {
         "servicePeriods": [
