--- modules/claims_api/spec/models/form_526_spec.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe ClaimsApi::Form526, type: :model do
-end
